added auton info
