+++ src/org/usfirst/frc/team1250/robot/AutoGroups/Auto_CenterPos.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team1250.robot.groups;
+package org.usfirst.frc.team1250.robot.AutoGroups;
 
 import org.usfirst.frc.team1250.robot.Robot;
 import org.usfirst.frc.team1250.robot.commandGroups.CmdG_AutonSwitch;
