+++ src/org/usfirst/frc/team1250/robot/AutoGroups/Auto_Fallback.java
@@ -1,0 +1,32 @@
+package org.usfirst.frc.team1250.robot.AutoGroups;
+
+import org.usfirst.frc.team1250.robot.elevator.Cmd_EleErectTower;
+
+import edu.wpi.first.wpilibj.command.CommandGroup;
+
+/**
+ *
+ */
+public class Auto_Fallback extends CommandGroup {
+
+    public Auto_Fallback() {
+    	
+    	addSequential(new Cmd_EleErectTower());
+        // Add Commands here:
+        // e.g. addSequential(new Command1());
+        //      addSequential(new Command2());
+        // these will run in order.
+
+        // To run multiple commands at the same time,
+        // use addParallel()
+        // e.g. addParallel(new Command1());
+        //      addSequential(new Command2());
+        // Command1 and Command2 will run in parallel.
+
+        // A command group will require all of the subsystems that each member
+        // would require.
+        // e.g. if Command1 requires chassis, and Command2 requires arm,
+        // a CommandGroup containing them would require both the chassis and the
+        // arm.
+    }
+}
