+++ src/org/usfirst/frc/team1250/robot/AutoGroups/Auto_LeftPos.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team1250.robot.groups;
+package org.usfirst.frc.team1250.robot.AutoGroups;
 
 import org.usfirst.frc.team1250.robot.Robot;
 import org.usfirst.frc.team1250.robot.commandGroups.CmdG_ScaleWhip;
@@ -6,36 +6,35 @@ import org.usfirst.frc.team1250.robot.commandGroups.CmdG_StriaghtScale;
 import org.usfirst.frc.team1250.robot.drive.Cmd_AutoDrive;
 import org.usfirst.frc.team1250.robot.drive.Cmd_AutoTurn;
 import org.usfirst.frc.team1250.robot.elevator.Cmd_EleErectTower;
-import org.usfirst.frc.team1250.robot.elevator.Cmd_EleSwitch;
 import org.usfirst.frc.team1250.robot.elevator.Cmd_EleHigh;
-
+import org.usfirst.frc.team1250.robot.elevator.Cmd_EleSwitch;
 import edu.wpi.first.wpilibj.command.CommandGroup;
 
 /**
  *
  */
-public class Auto_RightPos extends CommandGroup {
+public class Auto_LeftPos extends CommandGroup {
 
-    public Auto_RightPos() {
+    public Auto_LeftPos() {
     	addSequential(new Cmd_EleErectTower());
 // 		addSequential(new Cmd_AutoDrive(45));
     	
     	String autoMessage = Robot.getAutoMessage();
-
-    	if(autoMessage == "LR" || autoMessage =="RR")
+    	
+    	if(autoMessage == "RL" || autoMessage =="LL")
     	{
     		//Left Scale case
 //     		addParallel(new Cmd_AutoDrive(45));
 //     		addParallel(new Cmd_EleHigh());
 //     		addSequential(new Cmd_AutoTurn(90));
-    		addSequential(new CmdG_StriaghtScale(-1));
+    		addSequential(new CmdG_StriaghtScale(1));
     		
     	}
-    	else if(autoMessage == "RL" || autoMessage == "LL")
+    	else if(autoMessage == "LR" || autoMessage == "RR")
     	{
     		//Left Switch case
 //    		addParallel(new Cmd_AutoDrive(45));
-     		addSequential(new CmdG_ScaleWhip(-1));
+     		addSequential(new CmdG_ScaleWhip(1));
 //     		addSequential(new Cmd_AutoTurn(90));
     	}
     	else
